requirements.txt
Django
django-allauth
Pillow
matplotlib
rasterio
opencv-python
